+++ CHANGELOG.md
@@ -21,6 +21,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Removed check-quotes select menu as the maximum size is 25
 
+## [2.2.0]
+
+### Added
+
+- id parameter for command /check-quotes
+- /crawl-missing-quotes command to crawl the entire channel and find missing quotes
+
+### Changed
+
+- better random for quote selection
+
+### Fixed
+
+- Removed check-quotes select menu as the maximum size is 25
+
 ## [2.1.1]
 
 ### Added